feat!: gate sync client behind blocking feature; add tls features
